echo/tools/usage-tracker/pyproject.toml
[project]
name = "usage-tracker"
version = "0.1.0"
description = "Customer usage tracking and reporting tool for Dembrane ECHO"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.40.0",
    "litellm>=1.50.0",
    "requests>=2.32.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.2.0",
    "plotly>=5.24.0",
    "reportlab>=4.2.0",
    "wordcloud>=1.9.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "backoff>=2.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/usage_tracker"]

